rtp: basepay2: default to lower MTU of 1200 instead of 1400
.. for all payloaders but the MPEG-TS one where payloads of 1388 bytes
are typical, and which is primarily used on local networks anyway.

The lower MTU works better in scenarios where multiple layers of
encapsulation are involved such as VPNs or PPPoE, especially with
IPv6 in the mix.

Path MTU Discovery can't really be relied on in practice.

We've already been setting this in webrtcsink.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2785>